+++ packages/@averjs/cli/__fixtures__/EmptyCommand.js
@@ -1,3 +1,0 @@
-import Command from '../lib/commands/command';
-
-export default class EmptyCommand extends Command {}