+++ packages/@averjs/core/lib/global.ts
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-namespace */
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+declare namespace Express {
+  export interface Request {
+     id?: string;
+     error?: string;
+  }
+}